Object serialization to value tree supporting custom classes and arrays.
1. Multi-level object inspection resulting in object model (object description).
2. Object model serialization, object serialization based on model.
3. Deserialization.

+++ artemis-entity-tracker/src/main/java/net/namekdev/entity_tracker/network/communicator/ExternalInterfaceCommunicator.java
@@ -71,7 +71,7 @@ public class ExternalInterfaceCommunicator extends Communicator implements World
 					field.fieldName = _deserializer.readString();
 					field.classType = _deserializer.readString();
 					field.isArray = _deserializer.readBoolean();
-					field.valueType = _deserializer.readInt();
+					field.valueType = _deserializer.readByte();
 
 					info.fields.insertElementAt(field, i);
 				}
